refactor: change baseSalary type from BigDecimal to Long in User and UserEntity classes

+++ domain/model/src/main/java/co/com/crediya/model/user/User.java
@@ -22,6 +22,6 @@ public class User {
     private String address;
     private String phone;
     private String email;
-    private BigDecimal baseSalary;
+    private Long baseSalary;
     private String roleId;
 }
