+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/crediya/r2dbc/entity/UserEntity.java
@@ -23,6 +23,6 @@ public class UserEntity {
     private String address;
     private String phone;
     private String email;
-    private BigDecimal baseSalary;
+    private Long baseSalary;
     private String roleId;
 }
